conduit/requirements.txt
Package             Version
------------------- -------
asgiref             3.5.2
Django              4.1.3
djangorestframework 3.14.0
mysqlclient         2.1.1
pip                 22.2.2
pytz                2022.6
setuptools          63.2.0
sqlparse            0.4.3
tzdata              2022.6